Combined Prepare shared code, added limit variant.
