+++ src/prepare/index.js
@@ -3,3 +3,6 @@
  */
 export { default as webgl } from './webgl/WebGLPrepare';
 export { default as canvas } from './canvas/CanvasPrepare';
+export { default as BasePrepare } from './BasePrepare';
+export { default as CountLimiter } from './limiters/CountLimiter';
+export { default as TimeLimiter } from './limiters/TimeLimiter';
