+++ src/prepare/webgl/WebGLPrepare.js
@@ -1,7 +1,5 @@
 import * as core from '../../core';
-
-const SharedTicker = core.ticker.shared;
-const DEFAULT_UPLOADS_PER_FRAME = 4;
+import BasePrepare from '../BasePrepare';
 
 /**
  * The prepare manager provides functionality to upload content to the GPU.
@@ -9,239 +7,23 @@ const DEFAULT_UPLOADS_PER_FRAME = 4;
  * @class
  * @memberof PIXI
  */
-export default class WebGLPrepare
+export default class WebGLPrepare extends BasePrepare
 {
     /**
      * @param {PIXI.WebGLRenderer} renderer - A reference to the current renderer
      */
     constructor(renderer)
     {
-        /**
-         * Reference to the renderer.
-         * @type {PIXI.WebGLRenderer}
-         * @private
-         */
-        this.renderer = renderer;
+        super(renderer);
 
-        /**
-         * Collection of items to uploads at once.
-         * @type {Array<*>}
-         * @private
-         */
-        this.queue = [];
-
-        /**
-         * Collection of additional hooks for finding assets.
-         * @type {Array<Function>}
-         * @private
-         */
-        this.addHooks = [];
-
-        /**
-         * Collection of additional hooks for processing assets.
-         * @type {Array<Function>}
-         * @private
-         */
-        this.uploadHooks = [];
-
-        /**
-         * Callback to call after completed.
-         * @type {Array<Function>}
-         * @private
-         */
-        this.completes = [];
-
-        /**
-         * If prepare is ticking (running).
-         * @type {boolean}
-         * @private
-         */
-        this.ticking = false;
+        this.uploadHookSource = this.renderer;
 
         // Add textures and graphics to upload
         this.register(findBaseTextures, uploadBaseTextures)
             .register(findGraphics, uploadGraphics);
     }
 
-    /**
-     * Upload all the textures and graphics to the GPU.
-     *
-     * @param {Function|PIXI.DisplayObject|PIXI.Container} item - Either
-     *        the container or display object to search for items to upload or
-     *        the callback function, if items have been added using `prepare.add`.
-     * @param {Function} [done] - Optional callback when all queued uploads have completed
-     */
-    upload(item, done)
-    {
-        if (typeof item === 'function')
-        {
-            done = item;
-            item = null;
-        }
-
-        // If a display object, search for items
-        // that we could upload
-        if (item)
-        {
-            this.add(item);
-        }
-
-        // Get the items for upload from the display
-        if (this.queue.length)
-        {
-            this.numLeft = WebGLPrepare.UPLOADS_PER_FRAME;
-
-            if (done)
-            {
-                this.completes.push(done);
-            }
-
-            if (!this.ticking)
-            {
-                this.ticking = true;
-                SharedTicker.add(this.tick, this);
-            }
-        }
-        else if (done)
-        {
-            done();
-        }
-    }
-
-    /**
-     * Handle tick update.
-     *
-     * @private
-     */
-    tick()
-    {
-        // Upload the graphics
-        while (this.queue.length && this.numLeft > 0)
-        {
-            const item = this.queue[0];
-            let uploaded = false;
-
-            for (let i = 0, len = this.uploadHooks.length; i < len; i++)
-            {
-                if (this.uploadHooks[i](this.renderer, item))
-                {
-                    this.numLeft--;
-                    this.queue.shift();
-                    uploaded = true;
-                    break;
-                }
-            }
-
-            if (!uploaded)
-            {
-                this.queue.shift();
-            }
-        }
-
-        // We're finished
-        if (this.queue.length)
-        {
-            this.numLeft = WebGLPrepare.UPLOADS_PER_FRAME;
-        }
-        else
-        {
-            this.ticking = false;
-
-            SharedTicker.remove(this.tick, this);
-
-            const completes = this.completes.slice(0);
-
-            this.completes.length = 0;
-
-            for (let i = 0, len = completes.length; i < len; i++)
-            {
-                completes[i]();
-            }
-        }
-    }
-
-    /**
-     * Adds hooks for finding and uploading items.
-     *
-     * @param {Function} [addHook] - Function call that takes two parameters: `item:*, queue:Array`
-              function must return `true` if it was able to add item to the queue.
-     * @param {Function} [uploadHook] - Function call that takes two parameters: `renderer:WebGLRenderer, item:*` and
-     *        function must return `true` if it was able to handle upload of item.
-     * @return {PIXI.WebGLPrepare} Instance of plugin for chaining.
-     */
-    register(addHook, uploadHook)
-    {
-        if (addHook)
-        {
-            this.addHooks.push(addHook);
-        }
-
-        if (uploadHook)
-        {
-            this.uploadHooks.push(uploadHook);
-        }
-
-        return this;
-    }
-
-    /**
-     * Manually add an item to the uploading queue.
-     *
-     * @param {PIXI.DisplayObject|PIXI.Container|*} item - Object to add to the queue
-     * @return {PIXI.WebGLPrepare} Instance of plugin for chaining.
-     */
-    add(item)
-    {
-        // Add additional hooks for finding elements on special
-        // types of objects that
-        for (let i = 0, len = this.addHooks.length; i < len; i++)
-        {
-            if (this.addHooks[i](item, this.queue))
-            {
-                break;
-            }
-        }
-
-        // Get childen recursively
-        if (item instanceof core.Container)
-        {
-            for (let i = item.children.length - 1; i >= 0; i--)
-            {
-                this.add(item.children[i]);
-            }
-        }
-
-        return this;
-    }
-
-    /**
-     * Destroys the plugin, don't use after this.
-     *
-     */
-    destroy()
-    {
-        if (this.ticking)
-        {
-            SharedTicker.remove(this.tick, this);
-        }
-        this.ticking = false;
-        this.addHooks = null;
-        this.uploadHooks = null;
-        this.renderer = null;
-        this.completes = null;
-        this.queue = null;
-    }
-
 }
-
-/**
- * The number of graphics or textures to upload to the GPU
- *
- * @static
- * @type {number}
- * @default 4
- */
-WebGLPrepare.UPLOADS_PER_FRAME = DEFAULT_UPLOADS_PER_FRAME;
 
 /**
  * Built-in hook to upload PIXI.Texture objects to the GPU.
